Cleanup fixes.
Fixed remote vs local file deletion. Adaugat optiune "NODELETE" pentru
torrente catre TREBUIE sa ramana la seed.

+++ TransmissionAPIDIY/src/com/kiralycraft/tapidiy/Connection.java
@@ -28,6 +28,8 @@ public class Connection
 	
 	private JsonObject torrentInfo = null;
 	
+	boolean isLocalInstance;
+	
 	class TorrentInfo
 	{
 		private int id;
@@ -84,6 +86,30 @@ public class Connection
 		url+="/transmission/rpc";
 		this.url = url;
 		loginString = "Basic "+new String(encodedBytes);
+		
+		isLocalInstance = url.contains("localhost") || url.contains("127.0.0.1");
+		if (isLocalInstance)
+		{
+			System.out.println();
+			System.out.println();
+			System.out.println("===WARNING===");
+			System.out.println("Based on the Transmission URL, this is a LOCAL INSTANCE and files will be deleted accordingly.");
+			System.out.println("Please DO NOT \"localhost\" or \"127.0.0.1\" if this application is NOT RUNNING on the same machine as your Transmission instance.");
+			System.out.println("=============");
+			System.out.println();
+			System.out.println();
+		}
+		else
+		{
+			System.out.println();
+			System.out.println();
+			System.out.println("===WARNING===");
+			System.out.println("Based on the Transmission URL, this is an REMOTE INSTANCE and files will be deleted accordingly.");
+			System.out.println("Please use \"localhost\" or \"127.0.0.1\" if this application is running on the same machine as your Transmission instance.");
+			System.out.println("=============");
+			System.out.println();
+			System.out.println();
+		}
 	}
 	public int login()  
 	{
@@ -185,16 +211,19 @@ public class Connection
 	
 		            String name = torrentList.getAsJsonPrimitive("name").getAsString();
 		            
-		            long timeAddedSeconds = System.currentTimeMillis()/1000-torrentList.getAsJsonPrimitive("startDate").getAsLong();
-		            long lastActivitySeconds = System.currentTimeMillis()/1000 - torrentList.getAsJsonPrimitive("activityDate").getAsLong();
-		            
-		            TorrentInfo ti = new TorrentInfo(id,ratio,lastActivitySeconds,timeAddedSeconds,uploadEver,size,haveValid,name);
-		            
-		            tilist.add(ti); //doar daca indeplineste conditiile pentru a fi sters
-		            if ((ti.getUploadRatio()>1.0f || ti.getTimeAddedSeconds()>172800)) //daca are 48 de ore si nu o mai facut upload de o zi
+		            if (!name.contains("NODELETE"))
 		            {
-	//		            System.out.println(id+" "+ratio+" "+uploadEver+" "+size+" "+timeAddedSeconds+" "+lastActivitySeconds+" "+name);
-		            	ti.shouldBeProcessed=true;
+			            long timeAddedSeconds = System.currentTimeMillis()/1000-torrentList.getAsJsonPrimitive("startDate").getAsLong();
+			            long lastActivitySeconds = System.currentTimeMillis()/1000 - torrentList.getAsJsonPrimitive("activityDate").getAsLong();
+			            
+			            TorrentInfo ti = new TorrentInfo(id,ratio,lastActivitySeconds,timeAddedSeconds,uploadEver,size,haveValid,name);
+			            
+			            tilist.add(ti); //doar daca indeplineste conditiile pentru a fi sters
+			            if ((ti.getUploadRatio()>1.0f || ti.getTimeAddedSeconds()>172800)) //daca are 48 de ore si nu o mai facut upload de o zi
+			            {
+		//		            System.out.println(id+" "+ratio+" "+uploadEver+" "+size+" "+timeAddedSeconds+" "+lastActivitySeconds+" "+name);
+			            	ti.shouldBeProcessed=true;
+			            }
 		            }
 				}
 			}
@@ -211,8 +240,9 @@ public class Connection
 			////////////////////////////
 			////ACTUAL CLEANING/////////
 			////////////////////////////
+			long freeSpaceForDownloadingTorrents = getDownloadingTorrentsSpaceNeeded();  
 			Pair<Long, String> freespaceAndDownDir = getFreeSpaceAndDownDir();
-			long currentFreeSpace = freespaceAndDownDir.getKey();
+			long currentFreeSpace = freespaceAndDownDir.getKey()-freeSpaceForDownloadingTorrents;
 			String downloadDir = freespaceAndDownDir.getValue();
 			long freeableSpace = 0;
 			
@@ -242,17 +272,38 @@ public class Connection
 					for (TorrentInfo tmpti:toRemove)
 					{
 						System.out.println("Removing torrent instance from Transmission");
-						removeTorrent(tmpti.getId());
-						System.out.println("Removing associated files for:"+tmpti.getName());
-						JsonArray files = getTorrentFiles(tmpti.getId());
-						for (JsonElement je:files)
+						removeTorrent(tmpti.getId(),isLocalInstance);
+						if (isLocalInstance)
 						{
-							String toDeletePath = je.getAsJsonObject().getAsJsonPrimitive("name").getAsString();
-							File fileToDelete = new File(downloadDir+File.separator+toDeletePath);
-							System.out.println("Erasing "+toDeletePath);
-							if (!fileToDelete.delete())
+							System.out.println("Removing associated files for:"+tmpti.getName());
+							JsonArray files = getTorrentFiles(tmpti.getId());
+							for (JsonElement je:files)
 							{
-								System.out.println("Failed to delete!");
+								String toDeletePath = je.getAsJsonObject().getAsJsonPrimitive("name").getAsString();
+								File fileToDelete = new File(downloadDir+File.separator+toDeletePath);
+								System.out.println("Erasing "+toDeletePath);
+								
+								for (int i=0;i<=10;i++)
+								{
+									if (i==10)
+									{
+										System.out.println("Failed to delete after 10 tries! Aborting ...");
+										return false;
+									}
+									if (!fileToDelete.delete())
+									{
+										System.out.println("Failed to delete! Will retry in 1 second. Retry #"+i);
+										try
+										{
+											Thread.sleep(1000);
+										} 
+										catch (InterruptedException e)
+										{
+											;
+										}
+										
+									}
+								}
 							}
 						}
 					}
@@ -265,16 +316,11 @@ public class Connection
 			}
 		}
 	}
-//	public boolean removeTorrent(int id)
-//	{
-//		System.out.println(id);
-//		return true;
-//	}
-	public boolean removeTorrent(int id)
+	public boolean removeTorrent(int id, boolean localInstance)
 	{
 		try
 		{
-			byte[] postData       = ("{\"method\": \"torrent-remove\",\"arguments\": {\"delete-local-data\": false,\"ids\": ["+id+"]}}").getBytes( StandardCharsets.UTF_8 );
+			byte[] postData       = ("{\"method\": \"torrent-remove\",\"arguments\": {\"delete-local-data\": "+(!localInstance)+",\"ids\": ["+id+"]}}").getBytes( StandardCharsets.UTF_8 );
 			int    postDataLength = postData.length;
 			String request = url;
 			URL url = new URL(request);
